Time: 29 ms (87.01%) | Memory: 56.3 MB (91.47%) - LeetSync

+++ 1966-frequency-of-the-most-frequent-element/frequency-of-the-most-frequent-element.java
@@ -14,6 +14,6 @@ class Solution {
         res=Math.max(res,r-l+1);
             r+=1;
         }
-        return (int)res;
+        return res;
     }
 }
